Fix time mesurement

+++ com.Github.Haseoo.DASPP.CoreData/Dtos/ResultDto.cs
@@ -11,6 +11,6 @@ namespace com.Github.Haseoo.DASPP.CoreData.Dtos
         public int Vertex { get; set; }
 
         [JsonPropertyName("calculatingTimeMs")]
-        public int CalculatingTimeMs { get; set; }
+        public long CalculatingTimeMs { get; set; }
     }
 }
